+++ applications/fishing-map/src/features/timebar/timebar.selectors.ts
@@ -2,11 +2,11 @@ import { createSelector } from '@reduxjs/toolkit'
 import { DatasetTypes } from '@globalfishingwatch/api-types'
 import { selectTimebarGraph } from 'features/app/app.selectors'
 import {
-  resolveDataviewDatasetResource,
   selectActiveVesselsDataviews,
   selectEnvironmentalDataviews,
 } from 'features/workspace/workspace.selectors'
 import { selectResources, Resource, TrackResourceData } from 'features/resources/resources.slice'
+import { resolveDataviewDatasetResource } from 'features/resources/resources.selectors'
 
 type TimebarTrackSegment = {
   start: number