+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -38,6 +38,8 @@
 - Updated dependencies [ce98aa73]
   - @globalfishingwatch/layer-composer@4.17.0
   - @globalfishingwatch/react-hooks@4.9.4
+- Updated dependencies [[`d99bf5f3`](https://github.com/GlobalFishingWatch/frontend/commit/d99bf5f30bceed021ac990a0de8b4e992189be73)]:
+  - @globalfishingwatch/react-hooks@4.10.0
 
 ## 0.7.43