+++ packages/api-types/CHANGELOG.md
@@ -23,6 +23,7 @@
 ### Minor Changes
 
 - f9b53836: new dataview category type for events
+- [`d99bf5f3`](https://github.com/GlobalFishingWatch/frontend/commit/d99bf5f30bceed021ac990a0de8b4e992189be73) [#375](https://github.com/GlobalFishingWatch/frontend/pull/375) Thanks [@satellitestudiodesign](https://github.com/satellitestudiodesign)! - define endpoints enum
 
 ## 2.12.1