+++ packages/react-hooks/CHANGELOG.md
@@ -33,6 +33,7 @@
 
 - Updated dependencies [ce98aa73]
   - @globalfishingwatch/layer-composer@4.17.0
+- [`d99bf5f3`](https://github.com/GlobalFishingWatch/frontend/commit/d99bf5f30bceed021ac990a0de8b4e992189be73) [#375](https://github.com/GlobalFishingWatch/frontend/pull/375) Thanks [@satellitestudiodesign](https://github.com/satellitestudiodesign)! - define endpoints enum
 
 ## 4.9.3